backend/pyproject.toml
[project]
name = "django-nuxt-template"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [{ name = "hmeqo", email = "hmeqou@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "setuptools",
    "pydantic-settings",
    "django[argon2]",
    "django-cors-headers",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-simpleui",
    "django-cleanup",
    "django-dbbackup",
    "django-unused-media",
    "drf-apischema",
    "whitenoise[brotli]",
    "pillow",
    "platformdirs",
    "pendulum",
]

[dependency-groups]
types = [
    "django-stubs-ext",
    "django-stubs[compatible-mypy]",
    "django-types",
    "djangorestframework-stubs[compatible-mypy]",
]
cache = ["django-redis"]
postgresql = ["psycopg"]
mysql = ["mysqlclient"]
prod = ["granian"]
build = ["nuitka", "patchelf; platform_system == 'Linux'"]

[project.scripts]
dev = "project.scripts.dev:dev"
serve = "project.scripts.serve:serve"
build = "project.scripts.build:build"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apps", "src/backend", "src/project"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
extend-exclude = ["migrations"]
line-length = 120
ignore = ["F401", "F403", "F405"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "backend.settings"
